+++ NetEquus/Shared/Dtos/EstateDtos/EstateDto.cs
@@ -16,7 +16,7 @@ namespace Application.EstateApp.EstateDtos
 
         public int HorseCapacity { get; set; }
 
-        public int CurrentBalance { get; set; }
+        public decimal CurrentBalance { get; set; }
 
         public bool IsSytemEstate { get; set; }
     }
